+++ main.cpp
@@ -41,9 +41,9 @@ void make_direct_mapping_in_init_pml4() { // can be enabled only after init ppa
     uint64_t pages_to_map = calc_page_count_in_range(dm_start, dm_end, PAGE_SIZE::MB_2);
 
     for ( i = 0  ; i < pages_to_map ; i++ ) {
-        pd_offset = calc_pd_offset( i * PAGE_SIZE::MB_2 + dm_start);
-        pdpt_offset = calc_pdpt_offset(i * PAGE_SIZE::MB_2 + dm_start);
-        pml4_offset = calc_pml4_offset(i * PAGE_SIZE::MB_2 + dm_start);
+        pd_offset = calc_pd_offset( i * (uint64_t)PAGE_SIZE::MB_2 + dm_start);
+        pdpt_offset = calc_pdpt_offset(i * (uint64_t)PAGE_SIZE::MB_2 + dm_start);
+        pml4_offset = calc_pml4_offset(i * (uint64_t)PAGE_SIZE::MB_2 + dm_start);
 
         pd_for_dm[ pd_offset ] = ( vaddr_to_paddr_direct_mapping(i * 0x200000 + dm_start) & 0x000FFFFFFFFFF000 ) | 0x83 ;
         pdpt_for_hypervisor[ pdpt_offset ] = (reinterpret_cast<uint64_t>(&(pd_for_dm[ pd_offset ])) & 0x000FFFFFFFFFF000) | 0x23;
@@ -122,11 +122,11 @@ RootSystemDescriptionPointer* find_acpi_rsdp_bios() {
 
 void fill_hypervisor_final_vpt() {
     Address kernel_start_vaddr = reinterpret_cast<Address>(&_kernel_virtual_start);
-    Address kernel_end_vaddr = align_up(physical_page_allocator.get_page_array_end_vaddr(), PAGE_SIZE::MB_2);
+    Address kernel_end_vaddr = align_up(physical_page_allocator.get_page_array_end_vaddr(), (uint64_t)PAGE_SIZE::MB_2);
     uint64_t page_need_to_map = calc_page_count_in_range( kernel_start_vaddr, kernel_end_vaddr, PAGE_SIZE::MB_2);
 
     for ( auto i = 0 ; i < page_need_to_map ; i++ ) { // make kernel mapping
-        uint64_t current_vaddr = kernel_start_vaddr + i * PAGE_SIZE::MB_2;
+        uint64_t current_vaddr = kernel_start_vaddr + i * (uint64_t)PAGE_SIZE::MB_2;
         kernel_vpt.create_page_mapping( 
             current_vaddr,
             kernel_vaddr_to_paddr(current_vaddr),
@@ -136,11 +136,11 @@ void fill_hypervisor_final_vpt() {
     }
 
     Address direct_mapping_vstart = DIRECT_MAPPING_VSTART;
-    Address direct_mapping_vend = align_up( direct_mapping_vstart + physical_page_allocator.get_maximum_paddr(), PAGE_SIZE::MB_2 );
+    Address direct_mapping_vend = align_up( direct_mapping_vstart + physical_page_allocator.get_maximum_paddr(), (uint64_t)PAGE_SIZE::MB_2 );
     uint64_t dm_page_count = calc_page_count_in_range( direct_mapping_vstart, direct_mapping_vend, PAGE_SIZE::MB_2);
 
     for ( auto i = 0 ; i < dm_page_count ; i++ ) { // make direct mapping
-        uint64_t current_vaddr = direct_mapping_vstart + i * PAGE_SIZE::MB_2;
+        uint64_t current_vaddr = direct_mapping_vstart + i * (uint64_t)PAGE_SIZE::MB_2;
         kernel_vpt.create_page_mapping(
             current_vaddr,
             vaddr_to_paddr_direct_mapping( current_vaddr ),
